making duplicate now: problem is with async, tried init() method but it doesnt work
need to change so everytime there is a change, it is only to that user's branch. e.g. .child("recipe" + useremail without @).set rather than just fixed .child(recipe).set

after getting the duplicate set up, need to do it so when presses saved then it changes in that particular one and therefore shows up in personal + main while in main everything shows up.

duplicating it is more inefficient but is easier than creating a different type of list with different saveds and getting that.

//you have not saved anything! click on the search to find and save recipes

can you run code inside an if statement/ any statement? e.g. for loop inside if statement or:

reff.child("Recipe" + email.substring(0,(email.indexOf('@')))).setValue((for (int i = 0; i < 10; i++) { arr.add(0); }));

running the for loop inside the set value to get an answer

for loop as input for a method java

now that using seperate list of booleans with saved, need to have statements that take from recipe main and the one that shows and the one that takes from

have 3 sections, 1: edited, 1: main, 1: saved

when a new recipe is added to the main, then it just takes it from main.
edited is not shown in the main search area