user can select img now and it displays getting all data except for img and ingridients cant be added or removed, however existing ones can be added
// need to add functionality to add and remove ingridients
// need to add img functionality: push to firebase database

still need to add instructions however that is basically the same as ingridients without the checklist

now need to send to firebase
might want to download photos at the start to make it faster rather than downloading with picasso from the web using url everytime. however it might be too much effort.